Use PreOrderPartialChunkIterRef directly for selection
no more loop for subchunks...
