+++ src/tests.rs
@@ -11,10 +11,9 @@ use range_collections::{range_set::RangeSetEntry, RangeSet2, RangeSetRef};
 use crate::{
     blake3,
     io::{
-        fsm::{BaoContentItem, ResponseDecoderReadingNext},
         outboard::PreOrderMemOutboard,
         sync::{DecodeResponseItem, Outboard},
-        Header, Leaf, Parent,
+        Leaf,
     },
     iter::{
         encode_selected_rec, PostOrderChunkIter, PreOrderPartialChunkIterRef,
@@ -855,220 +854,6 @@ fn select_last_chunk_0() {
     assert_tuple_eq!(select_last_chunk_impl(1, 0));
 }
 
-/// Encode data fully, decode it again, and check that both data and outboard are the same
-///
-/// using the sync io api
-fn encode_decode_full_sync_impl(
-    data: &[u8],
-    outboard: PostOrderMemOutboard,
-) -> (
-    (Vec<u8>, PostOrderMemOutboard),
-    (Vec<u8>, PostOrderMemOutboard),
-) {
-    let ranges = RangeSet2::all();
-    let mut encoded = Vec::new();
-    crate::io::sync::encode_ranges_validated(&data, &outboard, &RangeSet2::all(), &mut encoded)
-        .unwrap();
-    let mut encoded_read = std::io::Cursor::new(encoded);
-    let mut decoded = Vec::new();
-    let ob_res_opt = crate::io::sync::decode_response_into(
-        outboard.root(),
-        outboard.tree().block_size,
-        &ranges,
-        &mut encoded_read,
-        |tree, root: blake3::Hash| {
-            let outboard_size = usize::try_from(tree.outboard_hash_pairs() * 64).unwrap();
-            let outboard_data = vec![0; outboard_size];
-            Ok(PostOrderMemOutboard::new(root, tree, outboard_data).unwrap())
-        },
-        &mut decoded,
-    )
-    .unwrap();
-    let ob_res = ob_res_opt.unwrap_or_else(|| {
-        PostOrderMemOutboard::new(outboard.root(), outboard.tree(), vec![]).unwrap()
-    });
-    ((decoded, ob_res), (data.to_vec(), outboard))
-}
-
-#[test]
-fn encode_decode_full_sync_cases() {
-    let cases = [(1024 + 1, 1)];
-    for (size, block_level) in cases {
-        let tree = BaoTree::new(ByteNum(size as u64), BlockSize(block_level));
-        let ranges = RangeSet2::all();
-        for item in ResponseIterRef::new(tree, &ranges) {
-            println!("ResponseChunk {:?}", item);
-        }
-        println!("\n");
-
-        let data = &make_test_data(size);
-        let block_size = BlockSize(block_level);
-        let outboard = PostOrderMemOutboard::create(&data, block_size);
-        let pair = encode_decode_full_sync_impl(data, outboard);
-        assert_tuple_eq!(pair);
-    }
-}
-
-fn encode_decode_partial_sync_impl(
-    data: &[u8],
-    outboard: PostOrderMemOutboard,
-    ranges: &RangeSetRef<ChunkNum>,
-) -> bool {
-    let mut encoded = Vec::new();
-    crate::io::sync::encode_ranges_validated(&data, &outboard, &ranges, &mut encoded).unwrap();
-    let expected_data = data;
-    let encoded_read = std::io::Cursor::new(encoded);
-    let buf = BytesMut::new();
-    let iter = crate::io::sync::DecodeResponseIter::new(
-        outboard.root,
-        outboard.tree.block_size,
-        encoded_read,
-        ranges,
-        buf,
-    );
-    for item in iter {
-        let item = match item {
-            Ok(item) => item,
-            Err(_) => {
-                return false;
-            }
-        };
-        match item {
-            DecodeResponseItem::Header(Header { size }) => {
-                // check that the size matches
-                if size != outboard.tree.size {
-                    return false;
-                }
-            }
-            DecodeResponseItem::Parent(Parent { node, pair }) => {
-                // check that the hash pair matches
-                if let Some(expected_pair) = outboard.load(node).unwrap() {
-                    if pair != expected_pair {
-                        return false;
-                    }
-                }
-            }
-            DecodeResponseItem::Leaf(Leaf { offset, data }) => {
-                // check that the data matches
-                if expected_data[offset.to_usize()..offset.to_usize() + data.len()] != data {
-                    return false;
-                }
-            }
-        }
-    }
-    true
-}
-
-/// Encode data fully, decode it again, and check that both data and outboard are the same
-///
-/// using the fsm io api
-async fn encode_decode_full_fsm_impl(
-    data: Vec<u8>,
-    outboard: PostOrderMemOutboard,
-) -> (
-    (Vec<u8>, PostOrderMemOutboard),
-    (Vec<u8>, PostOrderMemOutboard),
-) {
-    let mut outboard = outboard;
-    let ranges = RangeSet2::all();
-    let mut encoded = Vec::new();
-    crate::io::fsm::encode_ranges_validated(
-        Bytes::from(data.clone()),
-        &mut outboard,
-        &RangeSet2::all(),
-        &mut encoded,
-    )
-    .await
-    .unwrap();
-    let mut read_encoded = std::io::Cursor::new(encoded);
-    let mut decoded = BytesMut::new();
-    let ob_res_opt = crate::io::fsm::decode_response_into(
-        outboard.root(),
-        outboard.tree().block_size,
-        ranges,
-        &mut read_encoded,
-        |root, tree| async move {
-            let outboard_size = usize::try_from(tree.outboard_hash_pairs() * 64).unwrap();
-            let outboard_data = vec![0u8; outboard_size];
-            Ok(PostOrderMemOutboard::new(root, tree, outboard_data).unwrap())
-        },
-        &mut decoded,
-    )
-    .await
-    .unwrap();
-    let ob_res = ob_res_opt.unwrap_or_else(|| {
-        PostOrderMemOutboard::new(outboard.root(), outboard.tree(), vec![]).unwrap()
-    });
-    ((decoded.to_vec(), ob_res), (data, outboard))
-}
-
-async fn encode_decode_partial_fsm_impl(
-    data: &[u8],
-    outboard: PostOrderMemOutboard,
-    ranges: RangeSet2<ChunkNum>,
-) -> bool {
-    let mut encoded = Vec::new();
-    let mut outboard = outboard;
-    crate::io::fsm::encode_ranges_validated(
-        Bytes::from(data.to_vec()),
-        &mut outboard,
-        &ranges,
-        &mut encoded,
-    )
-    .await
-    .unwrap();
-    let expected_data = data;
-    let encoded_read = std::io::Cursor::new(encoded);
-    let initial = crate::io::fsm::ResponseDecoderStart::new(
-        outboard.root,
-        ranges,
-        outboard.tree.block_size,
-        encoded_read,
-    );
-    let (mut reading, size) = initial.next().await.unwrap();
-    if size != outboard.tree.size {
-        return false;
-    }
-    loop {
-        match reading.next().await {
-            ResponseDecoderReadingNext::More((reading1, result)) => {
-                let item = match result {
-                    Ok(item) => item,
-                    Err(_) => {
-                        return false;
-                    }
-                };
-                match item {
-                    BaoContentItem::Leaf(Leaf { offset, data }) => {
-                        // check that the data matches
-                        if expected_data[offset.to_usize()..offset.to_usize() + data.len()] != data
-                        {
-                            return false;
-                        }
-                    }
-                    BaoContentItem::Parent(Parent { node, pair }) => {
-                        // check that the hash pair matches
-                        if let Some(expected_pair) = outboard.load(node).unwrap() {
-                            if pair != expected_pair {
-                                return false;
-                            }
-                        }
-                    }
-                }
-                reading = reading1;
-            }
-            ResponseDecoderReadingNext::Done(_reader) => {
-                break;
-            }
-        }
-    }
-    true
-}
-
-fn block_size() -> impl Strategy<Value = BlockSize> {
-    (0..=6u8).prop_map(BlockSize)
-}
-
 /// Compares the PostOrderNodeIter with a simple stack-based reference implementation.
 #[test]
 fn test_post_order_node_iter() {
@@ -1323,42 +1108,6 @@ proptest! {
     #[test]
     fn pre_post_outboard(n in 0usize..1000000) {
         compare_pre_order_outboard(n);
-    }
-
-    #[test]
-    fn encode_decode_full_sync(size in 0usize..1<<17, block_size in block_size()) {
-        let data = make_test_data(size);
-        let outboard = PostOrderMemOutboard::create(&data, block_size);
-        let (expected, actual) = encode_decode_full_sync_impl(&data, outboard);
-        prop_assert_eq!(expected, actual);
-    }
-
-    #[test]
-    fn encode_decode_partial_sync((size, selection) in size_and_selection(0..100000, 2), block_size in block_size()) {
-        let data = make_test_data(size);
-        let outboard = PostOrderMemOutboard::create(&data, block_size);
-        let ok = encode_decode_partial_sync_impl(&data, outboard, &selection);
-        prop_assert!(ok);
-    }
-
-    #[test]
-    fn encode_decode_full_fsm(size in 0usize..1<<17, block_size in block_size()) {
-        let data = make_test_data(size);
-        let outboard = PostOrderMemOutboard::create(&data, block_size);
-        let (expected, actual) =
-            futures::executor::block_on(encode_decode_full_fsm_impl(data.into(), outboard))
-        ;
-        prop_assert_eq!(expected, actual);
-    }
-
-    #[test]
-    fn encode_decode_partial_fsm((size, selection) in size_and_selection(0..100000, 2), block_size in block_size()) {
-        let data = make_test_data(size);
-        let outboard = PostOrderMemOutboard::create(&data, block_size);
-        let ok =
-            futures::executor::block_on(encode_decode_partial_fsm_impl(&data, outboard, selection))
-        ;
-        prop_assert!(ok);
     }
 
     #[test]
